// src/types/FormData.ts
interface FormData {
    profession: string;
    name: string;
    lastName: string;
    location: any; // Use a location type for precise typing
    about: string;
    gender: string;
    interests: string[];
    favoriteColors: string[];
    place: string;
    latitude: string,
    longitude: string,

  }
  export default FormData;